Add files via upload
Creating a WhatsApp clone using HTML and CSS involves replicating the layout and basic design elements of the WhatsApp web interface. This guide will walk you through the steps to create a simple static clone of the WhatsApp chat interface.

1. Set Up the Project Structure
Create a new project folder and set up the basic file structure:  

2. Create the HTML Structure
In the index.html file, create the basic structure of the WhatsApp interface.

html


3. Style the Interface with CSS
In the styles.css file, add styles to replicate the WhatsApp design.